add command control and view
